对select进行了修改，使其可以接受一个字符串作为选择的内容，这样避免了只查询一个字段时还需要构造数组 对Manager进行了修改，现在Manager在进行增删改查前会对要执行操作的员工判断是否是自己所在部门的员工

+++ src/staff/Manager.java
@@ -1,5 +1,7 @@
 package staff;
 
+import java.util.HashMap;
+
 import db.*;
 
 public class Manager extends Staff{
@@ -9,20 +11,41 @@ public class Manager extends Staff{
 	public Manager(String staffID) {
 		super(staffID);
 	}
+	private boolean confirm(String staff_id) {
+		String thisSelectInfo[]= {"staff_id",this.staffID};
+		String SelectInfo[]= {"staff_id",staff_id};
+		String thisDpID=(String)select.selectSet(USER, PASS, "staff", "department_id", thisSelectInfo);
+		String staffDpID=(String)select.selectSet(USER, PASS, "staff", "department_id", SelectInfo);
+		if(thisDpID.equals(staffDpID))
+			return true;
+		return false;
+	}
 
 	//接受一个String数组，数组内容为员工的信息
-	public void addStaff(String[] column,String[] values) {
-		insert.insertSet(USER, PASS, "staff", column, values);
+	public boolean addStaff(String[] column,String[] values) {
+		if(confirm(values[0])) {
+			insert.insertSet(USER, PASS, "staff", column, values);
+			return true;
+		}
+		return false;
 	}
 	
 	//接受一个String，表示要删除的员工的员工号
-	public void delete(String staff_id) {
-		delete.deleteSet(USER, PASS, "staff", "staff_id", staff_id);
+	public boolean delete(String staff_id) {
+		if(confirm(staff_id)) {
+			delete.deleteSet(USER, PASS, "staff", "staff_id", staff_id);
+			return true;
+		}
+		return false;
 	}
 	
 	//传入两个个列表，第一个表示要修改的列，第二个表示修改内容
-	public void updateStaff(String[] column,String[] values,String staff_id) {
-		update.updateSet(USER, PASS, "staff", column, values, staff_id);
+	public boolean updateStaff(String[] column,String[] values,String staff_id) {
+		if(confirm(staff_id)) {
+			update.updateSet(USER, PASS, "staff", column, values, staff_id);
+			return true;
+		}
+		return false;
 	}
 	
 }
